Modify paddle.randint implementation to use int32->fp16 conversion
- Changed all paddle.randint calls to initialize with int32 dtype first
- Added .astype() conversion to target dtype (fp16, int8, etc.)
- Updated files:
  - utils/benchmark.py: Modified all benchmark functions
  - utils/paddle_package.py: Updated custom benchmark functions
  - src/triton/utils/quant/test.py: Fixed test tensor generation
  - bench/cuda/bench_qk_int8_pv_fp16_cuda.py: Updated CUDA benchmarks
  - bench/cuda/bench_qk_int8_pv_fp8_cuda.py: Updated FP8 benchmarks
- Added test script for verification
- Improves numerical stability and consistency

+++ bench/cuda/bench_qk_int8_pv_fp8_cuda.py
@@ -42,11 +42,11 @@ print(f"is_causal: {is_causal}")
 for seq_len in {128, 512, 1024, 2048}:
     flops = 4 * head * batch * headdim * seq_len * seq_len / (2 if is_causal else 1)
     q = paddle.randint(
-        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int8
-    ).cuda()
+        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int32
+    ).astype(paddle.int8).cuda()
     k = paddle.randint(
-        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int8
-    ).cuda()
+        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int32
+    ).astype(paddle.int8).cuda()
     o = paddle.empty(batch, seq_len, head, headdim, dtype=paddle.float16).cuda()
     vm = paddle.randn(batch, head, headdim, dtype=paddle.float32).cuda()
     v_scale = paddle.randn(batch, head, headdim, dtype=paddle.float32).cuda()
@@ -97,11 +97,11 @@ print(f"is_causal: {is_causal}")
 for seq_len in {128, 512, 1024, 2048}:
     flops = 4 * head * batch * headdim * seq_len * seq_len / (2 if is_causal else 1)
     q = paddle.randint(
-        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int8
-    ).cuda()
+        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int32
+    ).astype(paddle.int8).cuda()
     k = paddle.randint(
-        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int8
-    ).cuda()
+        low=-95, high=95, shape=(batch, seq_len, head, headdim), dtype=paddle.int32
+    ).astype(paddle.int8).cuda()
     o = paddle.empty(batch, seq_len, head, headdim, dtype=paddle.float16).cuda()
     vm = paddle.randn(batch, head, headdim, dtype=paddle.float32).cuda()
     v_scale = paddle.randn(batch, head, headdim, dtype=paddle.float32).cuda()
