+++ utils/benchmark.py
@@ -113,14 +113,14 @@ def benchmark_triton_attention_int8_baseline(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
@@ -216,14 +216,14 @@ def benchmark_triton_attention_int8(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         prepare_k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
@@ -319,14 +319,14 @@ def benchmark_triton_attention_int4(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         prepare_k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
@@ -429,14 +429,14 @@ def benchmark_triton_attention_q_int8_k_int4(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         prepare_k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
@@ -541,14 +541,14 @@ def benchmark_triton_attention_int2(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         prepare_k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
@@ -772,14 +772,14 @@ def benchmark_triton_attention_multi_kernel(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=q_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(q_dtype)
         k = paddle.randint(
             low=-100,
             high=100,
             shape=(batch_size, num_heads, seq_len, head_dim),
-            dtype=k_dtype,
-        )
+            dtype=paddle.int32,
+        ).astype(k_dtype)
         v = paddle.randn(
             batch_size, num_heads, seq_len, head_dim, dtype=v_dtype, device="cuda"
         )
